chore: update importer to only output unixfs things
BREAKING CHANGE:

If you specify `raw-nodes` to be true or `cid-version` to be `1`
and your data fits in one chunk, you would previously get a `CID`
that resolves to an `ipld-raw` node.

Now you will get a `CID` that resolves to a `dag-pb` node, so the
output of every import will be a `unixfs` object.